[IMP] rent.rent: cron job for deposit test

+++ rent/rent.py
@@ -1236,8 +1236,8 @@ class rent_rent(osv.osv):
 	def _check_deposit(self,cr,uid,args,context=None):
 		required_act = []
 		for record in args:
-			current = float(record['current_amount'])
-			depo = float(record['deposit'])
+			current = int(record['current_amount'])
+			depo = int(record['deposit'])
 			if current > depo:
 				required_act.append(rent_id)
 		debug(required_act)
